chore: move travellers file to actions
